Udacity reviewer fix: The other UILongPressGestureRecognizer states can also be utilized here. To create a drag effect you could do the following for each state:
On state .began , the pin will be created.
On state .changed , At this point you update the first pin coordinates to change the position of the pin
On state .ended , In this state the user has lifted the finger and you are free to persist the pin.
In addition the draggable property and mapView(_: annotationView: didChangeDragState: fromOldState: ) could be used.